+++ app/mailers/user_mailer.rb
@@ -99,7 +99,7 @@ class UserMailer < ApplicationMailer
     @recipient = recipient
     @email_to = recipient.email
     mail(to: recipient.email,
-         subject: "#{adverse_event.user.name} Reported a #{adverse_event.serious? ? 'Serious' : 'Non-Serious'} Adverse Event on #{adverse_event.project.name}",
+         subject: "#{adverse_event.user.name} Reported an Adverse Event on #{adverse_event.project.name}",
          reply_to: adverse_event.user.email)
   end