+++ db/migrate/20160125190315_remove_serious_from_adverse_events.rb
@@ -1,0 +1,5 @@
+class RemoveSeriousFromAdverseEvents < ActiveRecord::Migration
+  def change
+    remove_column :adverse_events, :serious, :boolean
+  end
+end